refactor(opsgenie): move configuration types into opsgenie package
