+++ config/notifiers_test.go
@@ -761,106 +761,6 @@ actions:
 	}
 }
 
-func TestOpsgenieTypeMatcher(t *testing.T) {
-	good := []string{"team", "user", "escalation", "schedule"}
-	for _, g := range good {
-		if !opsgenieTypeMatcher.MatchString(g) {
-			t.Fatalf("failed to match with %s", g)
-		}
-	}
-	bad := []string{"0user", "team1", "2escalation3", "sche4dule", "User", "TEAM"}
-	for _, b := range bad {
-		if opsgenieTypeMatcher.MatchString(b) {
-			t.Errorf("mistakenly match with %s", b)
-		}
-	}
-}
-
-func TestOpsGenieConfiguration(t *testing.T) {
-	for _, tc := range []struct {
-		name string
-		in   string
-
-		err bool
-	}{
-		{
-			name: "valid configuration",
-			in: `api_key: xyz
-responders:
-- id: foo
-  type: scheDule
-- name: bar
-  type: teams
-- username: fred
-  type: USER
-api_url: http://example.com
-`,
-		},
-		{
-			name: "api_key and api_key_file both defined",
-			in: `api_key: xyz
-api_key_file: xyz
-api_url: http://example.com
-`,
-			err: true,
-		},
-		{
-			name: "invalid responder type",
-			in: `api_key: xyz
-responders:
-- id: foo
-  type: wrong
-api_url: http://example.com
-`,
-			err: true,
-		},
-		{
-			name: "missing responder field",
-			in: `api_key: xyz
-responders:
-- type: schedule
-api_url: http://example.com
-`,
-			err: true,
-		},
-		{
-			name: "valid responder type template",
-			in: `api_key: xyz
-responders:
-- id: foo
-  type: "{{/* valid comment */}}team"
-api_url: http://example.com
-`,
-		},
-		{
-			name: "invalid responder type template",
-			in: `api_key: xyz
-responders:
-- id: foo
-  type: "{{/* invalid comment }}team"
-api_url: http://example.com
-`,
-			err: true,
-		},
-	} {
-		t.Run(tc.name, func(t *testing.T) {
-			var cfg OpsGenieConfig
-
-			err := yaml.UnmarshalStrict([]byte(tc.in), &cfg)
-			if tc.err {
-				if err == nil {
-					t.Fatalf("expected error but got none")
-				}
-				return
-			}
-
-			if err != nil {
-				t.Errorf("expected no error, got %v", err)
-			}
-		})
-	}
-}
-
 func TestSNS(t *testing.T) {
 	for _, tc := range []struct {
 		in  string
